Refactor the code: Move taskRunner to utils (BoincUtils)
