+++ android/BOINC/app/src/main/java/edu/berkeley/boinc/SplashActivity.java
@@ -29,17 +29,13 @@ import android.content.ServiceConnection;
 import android.graphics.Bitmap;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.Handler;
 import android.os.IBinder;
-import android.os.Looper;
 import android.os.RemoteException;
 import android.util.Log;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import java.util.concurrent.Callable;
-import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
 
 import edu.berkeley.boinc.attach.SelectionListActivity;
 import edu.berkeley.boinc.client.ClientStatus;
@@ -49,6 +45,7 @@ import edu.berkeley.boinc.databinding.ActivitySplashBinding;
 import edu.berkeley.boinc.ui.eventlog.EventLogActivity;
 import edu.berkeley.boinc.utils.BOINCUtils;
 import edu.berkeley.boinc.utils.Logging;
+import edu.berkeley.boinc.utils.TaskRunner;
 
 /**
  * Activity shown at start. Forwards to BOINCActivity automatically, once Monitor has connected to Client and received first data via RPCs.
@@ -94,29 +91,6 @@ public class SplashActivity extends AppCompatActivity {
             monitor = null;
         }
     };
-
-    public static class TaskRunner {
-        private final Executor executor = Executors.newSingleThreadExecutor(); // change according to your requirements
-        private final Handler handler = new Handler(Looper.getMainLooper());
-
-        public interface Callback<R> {
-            void onComplete(R result);
-        }
-
-        public <R> void executeAsync(Callable<R> callable, Callback<R> callback) {
-            executor.execute(() -> {
-                try {
-                    final R result = callable.call();
-                    handler.post(() -> {
-                        callback.onComplete(result);
-                    });
-                }catch(Exception e){
-                    Log.d(Logging.TAG, e.getMessage());
-                    e.printStackTrace();
-                }
-            });
-        }
-    }
 
     class BenchmarksTask implements Callable<Boolean> {
         @Override
